backend/pyproject.toml
[project]
name = "deriv-ai-assistant"
version = "0.1.0"
description = "For lablab.ai AI hackathon"
authors = [
    {name = "Eranda Sanjaya",email = "76166067+ES-3508@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pandas>=2.0.0",
    "google-generativeai>=0.3.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "google-genai (>=1.62.0,<2.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
